+++ src/CSVGMask.cpp
@@ -101,9 +101,7 @@ drawMask(const CSVGObject &object)
 
   //---
 
-  CMatrixStack2D transform;
-
-  svg_.getTransform(transform);
+  CMatrixStack2D transform = buffer->transform();
 
   // set transform
   CMatrixStack2D transform1;